+++ core/src/main/kotlin/cib/interns/test/task/core/controller/SocksMapper.kt
@@ -1,14 +1,14 @@
-package cib.interns.test.task.controller
+package cib.interns.test.task.core.controller
 
 import cib.interns.test.task.api.SocksIncomeRequest
 import cib.interns.test.task.api.SocksResponse
-import cib.interns.test.task.service.Socks
+import cib.interns.test.task.core.service.Socks
 import org.mapstruct.Mapper
 
 @Mapper(componentModel = "spring")
 interface SocksMapper {
 
-    fun transform(obj: SocksIncomeRequest):Socks
+    fun transform(obj: SocksIncomeRequest): Socks
 
     fun transform(obj: Socks):SocksResponse
 }
