+++ core/src/main/kotlin/cib/interns/test/task/core/service/SocksValidationDto.kt
@@ -1,4 +1,4 @@
-package cib.interns.test.task.service
+package cib.interns.test.task.core.service
 
 import javax.validation.constraints.Max
 import javax.validation.constraints.Min
